+++ src/components/Pages/footer/Footer.js
@@ -1,6 +1,6 @@
 
 import React from 'react';
-import './footer.css';
+import './Footer.css';
 
 function Footer() {
   return (
