+++ src/components/Pages/project/Project.js
@@ -16,7 +16,7 @@ import DeploymentModal from "../../Modal/DeploymentModal";
 import ReadOnlyEdgeModal from "../../Modal/ReadOnlyEdgeModal";
 import { useKeycloak } from "@react-keycloak/web";
 import { useParams } from "react-router-dom";
-import "./project.css";
+import "./Project.css";
 
 const readOnlyNodeStyle = {
   border: "1px solid #ccc",
